Split text formatting utilities and color utilities into separate files
