+++ lib/version.js
@@ -5,6 +5,7 @@
  *
  * For utilities specific to PHP version ranges, see php-version.js.
  */
+'use strict';
 
 const semver = require('semver');
 
@@ -28,21 +29,6 @@ function latest(versions) {
   return version;
 }
 exports.latest = latest;
-
-function color(version) {
-  var first = version[0];
-  if (first === 'v') {
-    first = version[1];
-  } else if (/^[0-9]/.test(version)) {
-    version = 'v' + version;
-  }
-  if (first === '0' || (version.indexOf('-') !== -1)) {
-    return { version: version, color: 'orange' };
-  } else {
-    return { version: version, color: 'blue' };
-  }
-}
-exports.color = color;
 
 function listCompare(a, b) {
   var alen = a.length, blen = b.length;
